backend/loginAPI/requirements.txt
Flask==3.1.0
psycopg2-binary
werkzeug
flask-cors